Fix ClassNotFoundException when using a mock inside a Spark job.
When running the tests with IntelliJ everything would run fine and the tests would pass. When running them with ./gradlew skyspark-scala:test they would fail with a ClassNotFoundException for the Mockito mocks that were used within a Spark job.

In the end it appears that running the tests with ./gradlew ended up with the mocks being serialized and deserialized across 2 different classloaders. Using the ACROSS_CLASSLOADERS Serializable mode fixes this.